User may create and delete question and answers

+++ app/controllers/questions_controller.rb
@@ -7,6 +7,8 @@ class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: %i[index show]
 
   def create
+    @exposed_question = current_user.questions.new(question_params)
+
     if question.save
       redirect_to question, notice: "Your question successfully created."
     else
@@ -23,8 +25,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: "Question was successfully deleted"
+    else
+      redirect_to questions_path, notice: "Can't delete the question"
+    end
   end
 
   private
